Use multiprocessing spawn method
Spawning the child process instead of forking it means we don't inherit
everything from the parent since spawn, unlike fork, won't actually copy
the parent process. We want this because using fork inherits the
singleton logging tree with all its handlers and they have various locks
that, if forked at the wrong time, will be held but won't be released in
the child process. spawn guarantees a fresh clean slate.

We set up our own context as not to mess around with anyone elses use of
the multiprocessing library within the same Python VM. Since we are
intended to be used as a library we really can't make any bets on how
the rest of the application wants to use the multiprocessing library and
thus a context is the sensible route to go.