Provide the worker process with a logger.
Now that we're spawning a new worker process instead of forking, the
logging magic autoconfiguration from the parent is also gone. If we want
to log from the child, we need to set up logging ourselves.

To simplify using this framework for the users, a logger will be
provided to the worker function automatically. The logger uses a
QueueHandler to send log messages from the child to the parent via a
queue. In the parent process, a QueueListener forwards the messages to
the logging handlers managed by NSO.