React to log level configuration changes
This adds yet another CDB subscriber to look for changes to the configured
python-vm logging level. Upon a change, it will signal the change to the child
process over a queue so that the child logger can be configured with the correct
logging level.

There was a choice here to implement the log level as a filter on the supervisor
side, so that the child process would always emit its log messages over the log
queue and the supervisor log queue listener could then drop those log messages
that didn't meet the correct log level. However, that means the child
continuously emits all log messages even when they are not needed. The standard
Python logger will upon being called, for example like log.debug(), first check
if any handler attached to the logging tree is interested in debug messages and
only if it finds one will it format the message and emit it. Since string
formatting is relatively expensive this means it can be skipped and this is what
makes it possible to use the logging library even in quite tight loops without a
big performance penalty (unless log level is set to debug, naturally). We want
to keep this promise of cheap logging, which is why we figured it was better to
propagate the log level to the child rather than implement a crude filter in the
parent supervisor.

This means that the bg function wrapper is setting up a thread in the child
process to listen to this. That's probably a little surprising for anyone that
wants to use this but weighing the benefits versus the drawbacks here, we
figured the promise of cheap logging is worth more. Users of this
background_process library should read up a bit on what it does anyway, or how
else would they use it?