Also react to vm-level logging level
The previous code only read and reacted to changes for the global python-vm
logging level. Apparently there is a per vm level as well. This new version will
react to changes to either or both values. It also correctly handles deletion of
the per vm value.

I am not very happy about the design here, I think there is room for improvement
but the design needs to be pondered for some time. This will do for now.