Use pipe to monitor child process liveness
This switches the liveness monitoring of the child process to use a pipe, which
is a sort of neat trick. This removes the last "busy polling" we had, even
though it wasn't very busy at one cycle per second, still this is more elegant.

A POSIX pipe will emit a EOF when the write end of the pipe is in the sole
possession of the child process and the child process dies. Since an EOF is a
waitable event we can wait for it using our select loop - the same select loop
we use for our main event queue.