React to HA changes
While the HA event notification listener has been in place for a while now we
haven't actually reacted to the messages that it writes to the supervisor queue.
This fixes that.

There is no way to configure the behaviour here. We will always just run the
worker when HA is disabled or when HA is enabled for the master node. There
might be scenarios where we want to run on slaves as well or only on slaves but
we'll complicate the code with such conditions when we need it.